Moved Tx button inside substrate/components folder
