+++ src/Upgrade.js
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
 import { Form, Input, Grid } from "semantic-ui-react";
 
-import TxButton from "./TxButton";
 import { useSubstrate } from "./substrate";
+import { TxButton } from "./substrate/components";
 
 export default function Upgrade(props) {
   const { api } = useSubstrate();
