+++ src/substrate/components/index.js
@@ -1,3 +1,4 @@
 import DeveloperConsole from "./DeveloperConsole";
+import TxButton from "./TxButton";
 
-export { DeveloperConsole };
+export { DeveloperConsole, TxButton };
